Refactors ion_assert's event stream comparison logic into the events library; provides a version of the library that links with gtest.
