+++ .gitignore
@@ -23,3 +23,4 @@ CPack*
 tools/cli/libion_cli.dylib
 tools/cli/ion
 tools/events/libion_events.dylib
+tools/events/libion_events_gtest.dylib
