+++ test/test_ion_binary.cpp
@@ -13,8 +13,10 @@
  */
 
 #include "ion_assert.h"
+#include "ion_event_stream.h"
 #include "ion_helpers.h"
 #include "ion_test_util.h"
+#include "ion_event_equivalence.h"
 
 TEST(IonBinaryLen, UInt64) {
