+++ test/test_ion_text.cpp
@@ -15,6 +15,7 @@
 #include "ion_assert.h"
 #include "ion_helpers.h"
 #include "ion_test_util.h"
+#include "ion_event_equivalence.h"
 
 TEST(IonTextSexp, ReaderHandlesNested)
 {
