+++ tools/events/ion_event_util.h
@@ -36,6 +36,9 @@
  */
 extern decContext g_IonEventDecimalContext;
 
+// Event stream marker
+static ION_STRING ion_cli_event_stream_symbol = {17, (BYTE *)"$ion_event_stream"}; // TODO where should this go?
+
 // Event fields
 static ION_STRING ion_event_event_type_field = {10, (BYTE *)"event_type"};
 static ION_STRING ion_event_ion_type_field = {8, (BYTE *)"ion_type"};
